Move optimized resources to resource package
