+++ src/ggrc/services/resources/assessment.py
